fix register page metadata error
